// frontend/src/lib/types.ts
export interface UserProfile {
    uid: string;
    name: string;
    email: string;
    profession: "student" | "professional" | "other" | "unspecified"; // Added 'unspecified' default
    address?: string;
    gender?: 'Male' | 'Female' | 'Rather not say';
    age?: number | null;
    emergencyNumber?: string;
    neurodiversity?: string[];
    createdAt: Date; // Keep the creation timestamp
}